complete task

+++ 0x06-pointers_arrays_strings/1-strncat.c
@@ -25,7 +25,7 @@ char *_strncat(char *dest, char *src, int n)
 		dest[i] = src[j];
 		i++;
 	}
-	if (strlen(src) < n)
+	if ((int) strlen(src) < n)
 	{
 		dest[i] = '\0';
 	}
